requirements.txt
Django~=3.0.4
celery~=4.4.2
kombu~=4.6.8
Pillow~=7.0.0
djangorestframework~=3.11.0
botocore~=1.15.32
dnspython~=1.16.0